% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{bar_graph}
\alias{bar_graph}
\title{A wrapper function for ggplot geom_bar}
\usage{
bar_graph(
  data,
  x,
  y = NA,
  z = NA,
  title = ggplot2::waiver(),
  subtitle = ggplot2::waiver(),
  color = NA
)
}
\arguments{
\item{data}{A list of data that ggplot can interpret.}

\item{x}{The explanatory variable in data.}

\item{y}{The response variable in data.}

\item{z}{A potential lurking variable in data.}

\item{title}{A string for the title.}

\item{subtitle}{A string for the subtitle.}

\item{color}{A literal color or a variable name from data.}
}
\value{
Returns a geom_bar plot from ggplot.
}
\description{
A wrapper function for ggplot geom_bar
}
\examples{
bar_graph(cereals, mfr, calories, title = 'Mfr vs Calories', 'mfr')
bar_graph(cereals, protein, title = 'Protein vs Calories', 'blue')

}
