% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{imports}
\alias{imports}
\title{Provides a template for the repetitive task of requiring packages It also installs missing packages.}
\usage{
imports(options = c("default"))
}
\arguments{
\item{options}{A vector of use cases. Choose from the following: default, csv, stat.tests, unit.tests}
}
\value{
No return value. Prints to the console upon package installation.
}
\description{
Provides a template for the repetitive task of requiring packages It also installs missing packages.
}
