% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{A wrapper function for ggplot geom_smooth}
\usage{
line_graph(
  data,
  x,
  y,
  z = NA,
  title = ggplot2::waiver(),
  subtitle = ggplot2::waiver(),
  one.chart = FALSE,
  show.CI = TRUE,
  color = NA
)
}
\arguments{
\item{data}{A list of data that ggplot can interpret.}

\item{x}{The explanatory variable in data.}

\item{y}{The response variable in data.}

\item{z}{A potential lurking variable in data.}

\item{title}{A string for the title.}

\item{subtitle}{A string for the subtitle.}

\item{one.chart}{A boolean to switch between single chart overlays or a facet wrap (default = FALSE).}

\item{show.CI}{A boolean to control the visibility of confidence intervals around line graphs (default = TRUE).}

\item{color}{A literal color or a variable name from data.}
}
\value{
Returns a geom_smooth plot from ggplot.
}
\description{
A wrapper function for ggplot geom_smooth
}
\examples{
line_graph(cereals, protein, calories, sugar, title = 'Is It Really Protein?', show.CI = FALSE, color = 'protein')
line_graph(cereals, sodium, fat, title = 'Fat vs Sodium', one.chart = TRUE, color = 'green')


}
