% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{A wrapper function for ggplot geom_point}
\usage{
scatter_plot(
  data,
  x,
  y,
  z = NA,
  title = ggplot2::waiver(),
  subtitle = ggplot2::waiver(),
  one.chart = FALSE,
  show.regression = FALSE,
  show.CI = TRUE,
  color = NA,
  size = NA
)
}
\arguments{
\item{data}{A list of data that ggplot can interpret.}

\item{x}{The explanatory variable in data.}

\item{y}{The response variable in data.}

\item{z}{A potential lurking variable in data.}

\item{title}{A string for the title.}

\item{subtitle}{A string for the subtitle.}

\item{one.chart}{A boolean to switch between single chart overlays or a facet wrap (default = FALSE).}

\item{show.regression}{A boolean to enable a regression line to appear (default = FALSE).}

\item{show.CI}{A boolean to control the visibility of confidence intervals around the regresion line (default = TRUE).}

\item{color}{A literal color or a variable from the data.}

\item{size}{Control the size of points through a variable name in data.}
}
\value{
Returns a geom_point plot from ggplot.
}
\description{
A wrapper function for ggplot geom_point
}
\examples{
scatter_plot(cereals, protein, sugar, title = 'Is It Really Protein?', show.regression = TRUE, color = 'calories')
line_graph(cereals, sugar, shelf, title = 'For Children?', show.regression = TRUE, show.CI = FALSE, size = 'protein')


}
